pyproject.toml
[tool.poetry]
name = "projeto-flask"
version = "0.1.0"
description = ""
authors = ["Carol <carolpsa.unesp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
pytz = "^2024.1"
flask-jwt-extended = "^4.6.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-ra -q -vvv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
